So i wasn't able to fix the issue what prevents me from rendering more than one text at a time

+++ SDLBreakout/SDLBreakout/Window.cpp
@@ -309,7 +309,7 @@ namespace Breakout {
         counted_frames++;
     }
     
-    float Window::get_fps()const{
+    double Window::get_fps()const{
         double fps = counted_frames / (timer.elapsed_time() / 1000.f);
         if(fps > 0xFFFFFFFF){
             fps = -1;
